Minor refactoring of syntax generator tests
Fix spelling mistakes
Use deconstructions on urlOperation to key, value paramaters
remove unnecessary async/await
remove unnecessary null-conditional operator
Fix incorect namespace compared to file placement
