+++ test/Atc.Rest.ApiGenerator.Tests/SyntaxGenerators/Api/ContractInterfaceTest.cs
@@ -9,7 +9,7 @@ using Microsoft.OpenApi.Models;
 using VerifyXunit;
 using Xunit;
 
-namespace Atc.Rest.ApiGenerator.Tests.SyntaxGenerators
+namespace Atc.Rest.ApiGenerator.Tests.SyntaxGenerators.Api
 {
     [UsesVerify]
     public class ContractInterfaceTest : SyntaxGeneratorTestBase
@@ -20,29 +20,29 @@ namespace Atc.Rest.ApiGenerator.Tests.SyntaxGenerators
 
         protected override ISyntaxCodeGenerator CreateApiGenerator(ApiProjectOptions apiProject)
         {
-            // Verify spec file suported for unit test
+            // Verify spec file supported for unit test
             Assert.Single(apiProject.Document.Paths);
             var urlPath = apiProject.Document.Paths.First();
-            Assert.False(urlPath.IsPathStartingSegmentName(FocusOnSecment));
+            Assert.False(urlPath.IsPathStartingSegmentName(FocusOnSegment));
             Assert.Single(urlPath.Value.Operations);
-            var urlOperation = urlPath.Value.Operations.First();
+            var (operationType, openApiOperation) = urlPath.Value.Operations.First();
 
             // Construct SUT
             return new SyntaxGeneratorContractInterface(
                         apiProject,
                         urlPath.Value.Parameters,
-                        urlOperation.Key,
-                        urlOperation.Value,
-                        FocusOnSecment,
-                        urlPath.Value.HasParameters() || urlOperation.Value.HasParametersOrRequestBody());
+                        operationType,
+                        openApiOperation,
+                        FocusOnSegment,
+                        urlPath.Value.HasParameters() || openApiOperation.HasParametersOrRequestBody());
         }
 
-        [Theory(DisplayName = "Contract Interface")]
+        [Theory(DisplayName = "Api Contract Interface")]
         [MemberData(nameof(YamlFiles))]
-        public async Task ExecuteGeneratorTest(YamlSpecFile specFile)
+        public Task ExecuteGeneratorTest(YamlSpecFile specFile)
         {
-            Assert.NotNull(specFile?.FilePath);
-            await ExecuteTest(specFile);
+            Assert.NotNull(specFile.FilePath);
+            return ExecuteTest(specFile);
         }
     }
 }
